fix panic when importing glh without the font file.
The old code was initializing the FontFile variable in the package's init
function. This was causing problems even when the client code wasn't
using the text rendering functions.

The old code was moved to the MustInitText function and the users of glh
now have the option to load, or not, the FontFile.

Users of glh should update the gas library since the search order was
changed to include the current folder too.